Rename conda_build_config.py to conda_build_config.yaml
Libsodium is not actually pinned in the current build, so when I upgraded zeromq it kept an old version of libsodium, resulting in shared library import errors. I think that the pin will get picked up if conda_build_config.py is renamed to conda_build_config.yaml. (A build number bump should accompany this.)